Move glibc-syscalls.h into bits/.
Bug: http://b/28519060
